+++ ui/src/components/select-list/QasSelectList.vue
@@ -2,17 +2,17 @@
   <qas-search-box v-bind="$attrs" class="q-pa-md" :fuse-options="fuseOptions" :list="sortedOptions">
     <template #default="{ results }">
       <q-list separator>
-        <q-item v-for="(result, index) in results" :key="index">
+        <q-item v-for="(item, index) in results" :key="index">
           <slot name="item" v-bind="self">
-            <slot name="item-section" :result="result">
+            <slot name="item-section" :result="item">
               <q-item-section class="items-start text-bold">
-                <div :class="labelClass" @click="redirectRoute(result)">{{ result.label }}</div>
+                <div :class="labelClass" @click="redirectRoute(item)">{{ item.label }}</div>
               </q-item-section>
             </slot>
 
             <q-item-section avatar>
               <slot name="item-action" v-bind="self">
-                <qas-btn hide-mobile-label v-bind="setButtonProps(result)" size="sm" @click="handleClick(result)" />
+                <qas-btn hide-mobile-label v-bind="setButtonProps(item)" size="sm" @click="handleClick(item)" />
               </slot>
             </q-item-section>
           </slot>